requirements.txt
geopandas==0.14.4
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pandas==2.2.2
shapely==2.0.4
matplotlib==3.9.0
seaborn==0.13.2
psutil==5.9.8
tqdm==4.66.4
python-dotenv==1.0.1
